requirements.txt
Django==2.1.7
django-filter==2.1.0
django-solo==1.1.3
django-tinymce4-lite==1.7.4
django-webpack-loader==0.6.0
djangorestframework==3.9.1
jsmin==2.2.2
Pillow==5.4.1
python-decouple==3.1
pytz==2018.9
